requirements.txt
flask
flask-cors
requests
python-dotenv
google-genai